+++ views/cameras-list/CameraTile.tsx
@@ -251,7 +251,11 @@ export const CameraTile: FC<CameraTileProps> = ({
               orientation: [
                 lockLandscapePlaybackOrientation ? 'sensorLandscape' : 'sensor',
               ],
+              backgroundColor: '#000000',
             },
+            topBar: {visible: false},
+            statusBar: {visible: false},
+            navigationBar: {visible: false, backgroundColor: '#000000'},
           },
         },
       });